refactor(utilities): add parse_telemetry_samples + relocate UTC helpers
Move _parse_srt_datetime/parse_utc_offset/estimate_utc_offset/resolve_utc_offset
to utilities (re-exported from telemetry_converter for back-compat) so the geo
Track layer can resolve UTC without a telemetry_converter -> geo.solar import
cycle. parse_telemetry_points is now a wrapper over parse_telemetry_samples.
